requirements.txt
annotated-types==0.6.0
anyio==4.1.0
black==23.11.0
certifi==2023.11.17
click==8.1.7
deepdiff==6.7.1
distro==1.8.0
exceptiongroup==1.2.0
faiss-cpu==1.9.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
langchain==0.3.7
langchain-community==0.3.5
langchain-openai==0.2.6
matplotlib==3.7.4
mypy-extensions==1.0.0
openai==1.54.3
packaging==23.2
pandas==2.0.3
pathspec==0.12.0
platformdirs==4.1.0
psutil==5.9.6
py-cpuinfo==9.0.0
python-dotenv==1.0.0
sniffio==1.3.0
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.12.2
scipy==1.14.1
cgroup_monitor==1.0.0
pydantic==2.10.6
